Rename package from redgirls to n2
All Java source files have been moved from the 'redgirls' package to the 'n2' package
