+++ src/main/java/n2/intellect/RedGirlsException.java
@@ -1,4 +1,4 @@
-package redgirls.intellect;
+package n2.intellect;
 
 public class RedGirlsException extends Exception {
     public RedGirlsException(String s) {
