+++ src/main/java/n2/purpose/EventTask.java
@@ -1,4 +1,4 @@
-package redgirls.purpose;
+package n2.purpose;
 
 public class EventTask extends Task {
     protected String from;
