+++ src/main/java/n2/purpose/TaskList.java
@@ -1,4 +1,4 @@
-package redgirls.purpose;
+package n2.purpose;
 
 import java.util.ArrayList;
