+++ text-ui-test/runtest.bat
@@ -7,15 +7,15 @@ REM delete output from previous run
 if exist ACTUAL.TXT del ACTUAL.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\redgirls\*.java
+javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\n2\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
 )
 REM no error here, errorlevel == 0
 
-REM run the program, feed redgirls.commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin redgirls.RedGirls < input.txt > ACTUAL.TXT
+REM run the program, feed n2.commands from input.txt file and redirect the output to the ACTUAL.TXT
+java -classpath ..\bin n2.RedGirls < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
