+++ text-ui-test/runtest.sh
@@ -19,7 +19,7 @@ then
     exit 1
 fi
 
-# run the program, feed redgirls.commands from input.txt file and redirect the output to the ACTUAL.TXT
+# run the program, feed n2.commands from input.txt file and redirect the output to the ACTUAL.TXT
 java -classpath ../bin RedGirl < input.txt > ACTUAL.TXT
 
 # convert to UNIX format
